[NEW] hicexplorer_hicrep.sh & TADs_loops_plots.R
- Addition of parameter "diffTADs" in hicexplorer_hicrep.sh and TADs_loops_plots.R.
In hicexplorer_hicrep.sh, if samples are all treated or all untreated, hicDifferentialTAD search for inputs, it doesn't find them and it stops without generating errors. No rejected/accepted files with TADs are generated and so no barplot and dataframe from TADs_loops_plots.R are created. Despite that, R script was trying to create them in any case, producing an error.

For this reason, now, if samples are NOT all treated or all untreated, diffTADs parameter will be equal to "true", hicDifferentialTAD will be executed and R script will creates the corresponding outputs. Otherwise not, but now it will not generates the error anymore.

hicDifferentialTAD looks for input files in any case. If it will not find them, it will not produces errors and hicexplorer_hicrep.sh will go on. To avoid the fact that it however search for its inputs file and to make things more clear, now differential TADs analysis chunk is moved in if-else statement, and executed only if diffTADs="true".

- TADs_loops_plots.R specific modification: now axis x test (samples lables) are removed because in case of samples with long name, those will be superimposed. Sample names are now available in the legend, which is now only one for all the plots, avoiding redudance.
- Height of the barplots are now reduced to better fit.